fetching digest from image field when using cri-o (PROJQUAY-6512) (#122)
As of OCP 4.15 the cri-o runtime now replaces the `pod.status.imageID` field with an internal cri-o ID instead of the resolved digest. This prevents CSO retrieving scan results since it's unable to fetch a digest for the resolved image. This change checks for a cri-o image ID and in which case will use the `pod.status.image` field. This allows images referenced by digest to be scanned. Images referenced by tag will not be able to be scanned.